// 3ME_UI/src/components/myPatients/patientAndTherapyDetails/woundProgress/addWoundAssessment/woundAssessmentPageSection.enum.ts
export enum WoundAssessmentPageSection {
  WOUND_ASSESSMENT_FORM = "woundassessmentform",
  WOUND_ASSESSMENT_REVIEW = "woundassessmentreview",
  WOUND_ASSESSMENT_SUMMARY = "woundassessmentresummary",
}
export enum WoundAssessmentType {
  MWP = "MWP",
  MWPWeekly = "MWPW",
  MWPBiWeekly = "MWPB",
  MWPFull = "MWPF",
}